+++ laktory/models/resources/databricks/job.py
@@ -295,6 +295,49 @@ class JobTaskPipelineTask(BaseModel):
     full_refresh: bool = None
 
 
+class JobTaskDbtTask(BaseModel):
+    """
+    dbt task specifications
+
+    Attributes
+    ----------
+    commands:
+        Series of dbt commands to execute in sequence. Every command must start with "dbt".
+    catalog:
+        The name of the catalog to use inside Unity Catalog.
+    profiles_directory:
+        The relative path to the directory in the repository specified by `git_source`
+        where dbt should look in for the `profiles.yml` file. If not specified, defaults
+        to the repository's root directory. Equivalent to passing `--profile-dir` to a
+        dbt command.
+    project_directory:
+        The path where dbt should look for dbt_project.yml. Equivalent to passing
+        `--project-dir` to the dbt CLI.
+        - If source is GIT: Relative path to the directory in the repository specified
+          in the git_source block. Defaults to the repository's root directory when not
+          specified.
+        - If source is WORKSPACE: Absolute path to the folder in the workspace.
+    schema:
+        The name of the schema dbt should run in. Defaults to `default`.
+    source:
+        The source of the project. Possible values are `WORKSPACE` and `GIT`. Defaults
+        to `GIT` if a `git_source` block is present in the job definition.
+    warehouse_id:
+        The ID of the SQL warehouse that dbt should execute against.
+
+        You also need to include a `git_source` block to configure the repository that
+        contains the dbt project.
+    """
+
+    commands: list[str]
+    catalog: str = None
+    profiles_directory: str = None
+    project_directory: str = None
+    schema: str = None
+    source: str = None
+    warehouse_id: str = None
+
+
 class JobTaskRunJobTask(BaseModel):
     """
     Job Task Run Job Task specifications
@@ -432,6 +475,8 @@ class JobTaskForEachTaskTask(BaseModel):
 
     Attributes
     ----------
+    dbt_task:
+        dbt task
     condition_task:
         Condition Task specifications
     depends_ons:
@@ -475,6 +520,7 @@ class JobTaskForEachTaskTask(BaseModel):
     """
 
     # compute_key: str = None
+    dbt_task: JobTaskDbtTask = None
     condition_task: JobTaskConditionTask = None
     depends_ons: list[JobTaskDependsOn] = None
     description: str = None