+++ laktory/models/datasinks/basedatasink.py
@@ -619,6 +619,7 @@ class BaseDataSink(BaseModel, PipelineChild):
         self,
         df: AnyDataFrame = None,
         mode: str = None,
+        full_refresh: bool = False,
         spark=None,
         view_definition: str = None,
     ) -> None:
@@ -631,6 +632,9 @@ class BaseDataSink(BaseModel, PipelineChild):
             Input dataframe.
         mode:
             Write mode overwrite of the sink default mode.
+        full_refresh
+            If `True`, tables are fully refreshed (re-built). Otherwise, only
+            increments are processed.
         spark:
             Spark Session for creating a view
         view_definition:
@@ -664,15 +668,19 @@ class BaseDataSink(BaseModel, PipelineChild):
             mode = self.mode
 
         if is_spark_dataframe(df):
-            self._write_spark(df=df, mode=mode)
+            self.dataframe_backend = "SPARK"
+            self._write_spark(df=df, mode=mode, full_refresh=full_refresh)
         elif is_polars_dataframe(df=df):
-            self._write_polars(df, mode=mode)
+            self.dataframe_backend = "POLARS"
+            self._write_polars(df, mode=mode, full_refresh=full_refresh)
         else:
             raise ValueError(f"DataFrame type '{type(df)}' not supported")
 
         logger.info("Write completed.")
 
-    def _write_spark(self, df: SparkDataFrame, mode: str = mode) -> None:
+    def _write_spark(
+        self, df: SparkDataFrame, mode: str = mode, full_refresh: bool = False
+    ) -> None:
         raise NotImplementedError("Not implemented for Spark dataframe")
 
     def _write_spark_view(self, view_definition: str, spark) -> None:
@@ -680,12 +688,22 @@ class BaseDataSink(BaseModel, PipelineChild):
             f"View creation with spark is not implemented for type '{type(self)}'"
         )
 
-    def _write_polars(self, df: PolarsLazyFrame, mode=mode) -> None:
+    def _write_polars(
+        self, df: PolarsLazyFrame, mode=mode, full_refresh: bool = False
+    ) -> None:
         raise NotImplementedError("Not implemented for Polars dataframe")
 
     # ----------------------------------------------------------------------- #
     # Purge                                                                   #
     # ----------------------------------------------------------------------- #
+
+    def exists(self, spark=None):
+        try:
+            df = self.read(spark=spark, as_stream=False)
+            df.limit(1).collect()
+            return True
+        except Exception:
+            return False
 
     def _purge_checkpoint(self, spark=None):
         if self._checkpoint_location:
@@ -735,6 +753,10 @@ class BaseDataSink(BaseModel, PipelineChild):
         """
         Delete sink data and checkpoints
         """
+        # TODO: Now that sink switch to overwrite when sink does not exists or when
+        # a full refresh is requested, the purge method should not delete the data
+        # by default, but only the checkpoints. Also consider truncating the table
+        # instead of dropping it.
         raise NotImplementedError()
 
     # ----------------------------------------------------------------------- #