+++ laktory/models/datasinks/tabledatasink.py
@@ -153,7 +153,7 @@ class TableDataSink(BaseDataSink):
     # Methods                                                                 #
     # ----------------------------------------------------------------------- #
 
-    def _write_spark(self, df: SparkDataFrame, mode=None) -> None:
+    def _write_spark(self, df: SparkDataFrame, mode=None, full_refresh=False) -> None:
         if df.isStreaming and self._checkpoint_location is None:
             raise ValueError("Checkpoint must be provided for streaming table sink.")
 
@@ -161,7 +161,9 @@ class TableDataSink(BaseDataSink):
             mode = self.mode
 
         if self.warehouse == "DATABRICKS":
-            return self._write_spark_databricks(df, mode=mode)
+            return self._write_spark_databricks(
+                df, mode=mode, full_refresh=full_refresh
+            )
         else:
             raise NotImplementedError(
                 f"Warehouse '{self.warehouse}' is not yet supported."
@@ -173,13 +175,29 @@ class TableDataSink(BaseDataSink):
         if self.parent_pipeline_node:
             self.parent_pipeline_node._output_df = df
 
-    def _write_spark_databricks(self, df: SparkDataFrame, mode) -> None:
+    def _write_spark_databricks(
+        self, df: SparkDataFrame, mode, full_refresh=False
+    ) -> None:
         if self.format in ["EXCEL"]:
             raise ValueError(f"'{self.format}' format is not supported with Spark")
 
         if mode.lower() == "merge":
             self.merge_cdc_options.execute(source=df)
             return
+
+        # Full Refresh
+        if full_refresh or not self.exists(spark=df.sparkSession):
+            if df.isStreaming:
+                if df.laktory.is_aggregate():
+                    logger.info(
+                        "Full refresh or initial load. Switching to COMPLETE mode."
+                    )
+                    mode = "COMPLETE"
+            else:
+                logger.info(
+                    "Full refresh or initial load. Switching to OVERWRITE mode."
+                )
+                mode = "OVERWRITE"
 
         # Default Options
         _options = {"mergeSchema": "true", "overwriteSchema": "false"}
@@ -225,6 +243,11 @@ class TableDataSink(BaseDataSink):
         """
         Delete sink data and checkpoints
         """
+        # TODO: Now that sink switch to overwrite when sink does not exists or when
+        # a full refresh is requested, the purge method should not delete the data
+        # by default, but only the checkpoints. Also consider truncating the table
+        # instead of dropping it.
+
         # Remove Data
         if self.warehouse == "DATABRICKS":
             logger.info(
@@ -277,6 +300,8 @@ class TableDataSink(BaseDataSink):
         if as_stream:
             source.as_stream = as_stream
 
+        if self.dataframe_backend:
+            source.dataframe_backend = self.dataframe_backend
         source.parent = self.parent
 
         return source