+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 * n/a
 ### Fixed
-* n/a
+* SCD Type 2 Sink without delete condition caused exception [[#361](https://github.com/okube-ai/laktory/issues/361)]
 ### Updated
 * n/a
 ### Breaking changes