+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 ### Fixed
 * SCD Type 2 Sink without delete condition caused exception [[#361](https://github.com/okube-ai/laktory/issues/361)]
 ### Updated
-* Validation for self referencing variables [[#362](https://github.com/okube-ai/laktory/issues/362)]
+* n/a
+### Breaking changes
+* n/a
 
 ## [0.6.4] - 2025-01-24
 ### Fixed