+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## [0.6.5] - Unreleased
+### Added
+* dbt job task
 ### Fixed
 * SCD Type 2 Sink without delete condition caused exception [[#361](https://github.com/okube-ai/laktory/issues/361)]
 ### Updated